♻️(front) move upload component under form/file-uploader
The uploader is a form control, so it now lives alongside the other
form components and its stories under Components/Forms. The move also
brings an MDX docs page. Root exports are unchanged for consumers.
