+++ e2e/helpers/mount-upload.tsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
 import { Modal, ModalSize } from "@gouvfr-lasuite/cunningham-react";
 import { CunninghamProvider } from "../../src/components/Provider/Provider";
-import { FileUploader } from "../../src/components/upload/FileUploader";
-import { UploadFile } from "../../src/components/upload/types";
+import { FileUploader } from "../../src/components/form/file-uploader/FileUploader";
+import { UploadFile } from "../../src/components/form/file-uploader/types";
 
 const GB = 1000 * 1000 * 1000;
