+++ src/components/form/file-uploader/FileUploader.mdx
@@ -1,0 +1,126 @@
+import { Meta, Canvas, ArgTypes } from "@storybook/blocks";
+import * as FileUploaderStories from "./file-uploader.stories";
+
+<Meta of={FileUploaderStories} name="Docs" />
+
+# FileUploader
+
+`FileUploader` is a drop zone where users can drag & drop files or browse the file system. It is fully controlled: you own the list of files and their upload states, and the component renders them.
+
+Labels are translated via Cunningham, so the component must live under a `CunninghamProvider`. The styles are bundled in `@gouvfr-lasuite/ui-kit/style`, so make sure it is imported once in your app.
+
+## Usage
+
+```tsx
+import { FileUploader, UploadFile } from "@gouvfr-lasuite/ui-kit";
+
+const [files, setFiles] = useState<UploadFile[]>([]);
+
+<FileUploader
+  files={files}
+  maxSize={5 * 1000 * 1000 * 1000}
+  onAddFiles={setFiles}
+  onRemoveFile={(file) =>
+    setFiles((prev) => prev.filter((f) => f.id !== file.id))
+  }
+/>;
+```
+
+<Canvas of={FileUploaderStories.Empty} />
+
+## The controlled model
+
+The component never keeps files internally. When the user selects or drops files, `onAddFiles` is called with the **next** collection: the replacement file in single mode, or the current `files` plus the added ones in multiple mode. Each entry is an `UploadFile`:
+
+```ts
+type UploadFile = {
+  /** Stable identifier. */
+  id: string;
+  /** Native file selected by the user, including its binary contents. */
+  originalFile: File;
+  /** "pending" | "uploading" | "done" (default) | "error" */
+  status?: UploadFileStatus;
+  /** Upload progress percentage, from 0 to 100, used when status is "uploading". */
+  progress?: number;
+  /** Short error message shown inline when status is "error". */
+  error?: string;
+  /** Optional detailed error message shown when hovering or focusing the error icon. */
+  errorDetails?: string;
+};
+```
+
+Start your upload from `onAddFiles`, then update each file's `status` and `progress` as the request advances — the uploader re-renders accordingly.
+
+## Single or multiple files
+
+By default only one file can be selected; the dropzone itself displays the file and its state (uploading / done / error), and picking a new file replaces it:
+
+<Canvas of={FileUploaderStories.SingleDone} />
+
+Pass `multiple` to accept several files: the dropzone stays available and the files are listed below it, each row with its own progress, error, and remove/cancel controls:
+
+<Canvas of={FileUploaderStories.MultipleStates} />
+
+<Canvas of={FileUploaderStories.InteractiveMultiple} />
+
+The component forwards `accept` to the native `<input type="file">`, so file types can be restricted:
+
+```tsx
+<FileUploader multiple accept="image/png, image/jpeg" />
+```
+
+`maxSize` (in bytes) is displayed as a "Max …" hint in the dropzone. It is informative only: enforce the limit yourself and flag oversized files with `status: "error"`.
+
+## States
+
+A file with `status: "uploading"` shows a spinner (single mode) or a progress bar (multiple mode); `"error"` shows the `error` message — plus `errorDetails` on the error icon's tooltip in multiple mode:
+
+<Canvas of={FileUploaderStories.SingleUploading} />
+
+<Canvas of={FileUploaderStories.SingleError} />
+
+A typical flow cycles through those states:
+
+<Canvas of={FileUploaderStories.SimulatedUpload} />
+
+## Remove and cancel
+
+The remove control appears on done or errored files only when `onRemoveFile` is provided; the cancel control appears on uploading files only when `onCancelFile` is provided. Both are called with the targeted `UploadFile` — filtering it out of your state removes the row:
+
+```tsx
+<FileUploader
+  files={files}
+  onAddFiles={setFiles}
+  onRemoveFile={(file) =>
+    setFiles((prev) => prev.filter((f) => f.id !== file.id))
+  }
+  onCancelFile={(file) => {
+    abortUpload(file.id);
+    setFiles((prev) => prev.filter((f) => f.id !== file.id));
+  }}
+/>
+```
+
+## In a modal
+
+The uploader adapts to its container width, e.g. inside a `Modal`:
+
+<Canvas of={FileUploaderStories.InModal} />
+
+## Custom labels
+
+Every text is translated via Cunningham (the `components.upload.*` keys) and can be overridden per instance through `labels`:
+
+```tsx
+<FileUploader
+  labels={{
+    noFileYet: "Drop your export here",
+    clickToUpload: "Browse…",
+    maxSize: "Max 2 MB per file",
+  }}
+/>
+```
+
+## Props
+
+<ArgTypes of={FileUploaderStories} />
