+++ src/components/form/index.tsx
@@ -1,1 +1,2 @@
 export * from "./label";
+export * from "./file-uploader";
